refactor(): more interdepency solving: - Plane does not depend on Polygon3 - move splitPolygon to tree.js module, renamed to splitPolygonByPlane - moved fromPolygons back to CSG.js ...again
