+++ src/math/Plane.js
@@ -1,6 +1,4 @@
 const Vector3D = require('./Vector3')
-const Vertex = require('./Vertex3')
-const Polygon = require('./Polygon3')
 const Line3D = require('./Line3')
 const {EPS, getTag} = require('../constants')
 
@@ -10,8 +8,6 @@ const Plane = function (normal, w) {
   this.normal = normal
   this.w = w
 }
-
-module.exports = Plane
 
 // create from an untyped object with identical property names:
 Plane.fromObject = function (obj) {
@@ -103,120 +99,6 @@ Plane.prototype = {
     return newplane
   },
 
-    // Returns object:
-    // .type:
-    //   0: coplanar-front
-    //   1: coplanar-back
-    //   2: front
-    //   3: back
-    //   4: spanning
-    // In case the polygon is spanning, returns:
-    // .front: a Polygon of the front part
-    // .back: a Polygon of the back part
-  splitPolygon: function (polygon) {
-    let result = {
-      type: null,
-      front: null,
-      back: null
-    }
-        // cache in local lets (speedup):
-    let planenormal = this.normal
-    let vertices = polygon.vertices
-    let numvertices = vertices.length
-    if (polygon.plane.equals(this)) {
-      result.type = 0
-    } else {
-      let thisw = this.w
-      let hasfront = false
-      let hasback = false
-      let vertexIsBack = []
-      let MINEPS = -EPS
-      for (let i = 0; i < numvertices; i++) {
-        let t = planenormal.dot(vertices[i].pos) - thisw
-        let isback = (t < 0)
-        vertexIsBack.push(isback)
-        if (t > EPS) hasfront = true
-        if (t < MINEPS) hasback = true
-      }
-      if ((!hasfront) && (!hasback)) {
-                // all points coplanar
-        let t = planenormal.dot(polygon.plane.normal)
-        result.type = (t >= 0) ? 0 : 1
-      } else if (!hasback) {
-        result.type = 2
-      } else if (!hasfront) {
-        result.type = 3
-      } else {
-                // spanning
-        result.type = 4
-        let frontvertices = []
-        let backvertices = []
-        let isback = vertexIsBack[0]
-        for (let vertexindex = 0; vertexindex < numvertices; vertexindex++) {
-          let vertex = vertices[vertexindex]
-          let nextvertexindex = vertexindex + 1
-          if (nextvertexindex >= numvertices) nextvertexindex = 0
-          let nextisback = vertexIsBack[nextvertexindex]
-          if (isback === nextisback) {
-                        // line segment is on one side of the plane:
-            if (isback) {
-              backvertices.push(vertex)
-            } else {
-              frontvertices.push(vertex)
-            }
-          } else {
-                        // line segment intersects plane:
-            let point = vertex.pos
-            let nextpoint = vertices[nextvertexindex].pos
-            let intersectionpoint = this.splitLineBetweenPoints(point, nextpoint)
-            let intersectionvertex = new Vertex(intersectionpoint)
-            if (isback) {
-              backvertices.push(vertex)
-              backvertices.push(intersectionvertex)
-              frontvertices.push(intersectionvertex)
-            } else {
-              frontvertices.push(vertex)
-              frontvertices.push(intersectionvertex)
-              backvertices.push(intersectionvertex)
-            }
-          }
-          isback = nextisback
-        } // for vertexindex
-                // remove duplicate vertices:
-        let EPS_SQUARED = EPS * EPS
-        if (backvertices.length >= 3) {
-          let prevvertex = backvertices[backvertices.length - 1]
-          for (let vertexindex = 0; vertexindex < backvertices.length; vertexindex++) {
-            let vertex = backvertices[vertexindex]
-            if (vertex.pos.distanceToSquared(prevvertex.pos) < EPS_SQUARED) {
-              backvertices.splice(vertexindex, 1)
-              vertexindex--
-            }
-            prevvertex = vertex
-          }
-        }
-        if (frontvertices.length >= 3) {
-          let prevvertex = frontvertices[frontvertices.length - 1]
-          for (let vertexindex = 0; vertexindex < frontvertices.length; vertexindex++) {
-            let vertex = frontvertices[vertexindex]
-            if (vertex.pos.distanceToSquared(prevvertex.pos) < EPS_SQUARED) {
-              frontvertices.splice(vertexindex, 1)
-              vertexindex--
-            }
-            prevvertex = vertex
-          }
-        }
-        if (frontvertices.length >= 3) {
-          result.front = new Polygon(frontvertices, polygon.shared, polygon.plane)
-        }
-        if (backvertices.length >= 3) {
-          result.back = new Polygon(backvertices, polygon.shared, polygon.plane)
-        }
-      }
-    }
-    return result
-  },
-
     // robust splitting of a line by a plane
     // will work even if the line is parallel to the plane
   splitLineBetweenPoints: function (p1, p2) {
@@ -254,3 +136,5 @@ Plane.prototype = {
     return mirrored
   }
 }
+
+module.exports = Plane
